Adds implementation for OAuth providers
This commit is backwards compatible because if a "provider" property is not available, it assumes the password login. 
If there is an "provider" property on the options hash (that is not password), then it utilizes Firebase's OAuth login.
An additional property of "redirect" is available as a boolean to allow users to pick between redirect and popup OAuth login.